+++ validator/src/main.rs
@@ -1733,8 +1733,8 @@ pub fn main() {
         let mut archive_format = ArchiveFormat::from_cli_arg(&archive_format_str)
             .unwrap_or_else(|| panic!("Archive format not recognized: {archive_format_str}"));
         if let ArchiveFormat::TarZstd { config } = &mut archive_format {
-            // level 1 is optimized for speed
-            config.compression_level = 1;
+            config.compression_level =
+                value_t_or_exit!(matches, "snapshot_zstd_compression_level", i32);
         }
         archive_format
     };